Add files via upload
Perform data cleaning on a real-world data set of artworks contained in the Museum of Modern Art (MoMA). MoMA, a museum in New York City, has one of the largest collections of modern art in the world.
In this exercise, we parse numbers from complex string and exercise differnt methods in manupuating strings.